refactor(shared): extract reusable domain preconditions (#14)
