+++ src/main/java/br/com/f2e/ovenplatform/shared/domain/validation/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("validation")
+package br.com.f2e.ovenplatform.shared.domain.validation;
